+++ src/components/panelContents/DetailPanel.vue
@@ -52,63 +52,11 @@ Copyright (C) 2021, Tomoya Komiyama.
       <div class="field">
         <CheckboxInput v-model="inheritScale" label="Inherit Scale" />
       </div>
-      <div class="field inline">
-        <label>Constraints</label>
-        <SelectButton
-          :options="constraintOptions"
-          @select="setBoneConstraintName"
-        />
-      </div>
-      <div
-        v-for="(c, i) in lastSelectedBone.constraints"
-        :key="i"
-        class="constraints-item"
-      >
-        <template v-if="c.name === 'IK'">
-          <IKOptionField
-            :model-value="c.option"
-            :bone-options="otherBoneOptions"
-            @update:modelValue="(option) => updateConstraint(i, option)"
-          />
-        </template>
-        <template v-else-if="c.name === 'LIMIT_ROTATION'">
-          <LimitRotationOptionField
-            :model-value="c.option"
-            @update:modelValue="(option) => updateConstraint(i, option)"
-          />
-        </template>
-        <template v-else-if="c.name === 'COPY_LOCATION'">
-          <CopyLocationOptionField
-            :model-value="c.option"
-            :bone-options="otherBoneOptions"
-            @update:modelValue="(option) => updateConstraint(i, option)"
-          />
-        </template>
-        <template v-else-if="c.name === 'COPY_ROTATION'">
-          <CopyRotationOptionField
-            :model-value="c.option"
-            :bone-options="otherBoneOptions"
-            @update:modelValue="(option) => updateConstraint(i, option)"
-          />
-        </template>
-        <div class="constraint-buttons">
-          <button :disabled="i === 0" type="button" @click="upConstraint(i)">
-            <UpIcon class="icon" />
-          </button>
-          <button
-            :disabled="
-              !lastSelectedBone || i === lastSelectedBone.constraints.length - 1
-            "
-            type="button"
-            @click="downConstraint(i)"
-          >
-            <UpIcon class="icon" flipped />
-          </button>
-          <button type="button" @click="deleteConstraint(i)">
-            <DeleteIcon class="icon" />
-          </button>
-        </div>
-      </div>
+      <ConstraintList
+        :constraints="lastSelectedBone.constraints"
+        :bone-options="otherBoneOptions"
+        @update="updateConstraints"
+      />
     </form>
   </div>
 </template>
@@ -118,31 +66,15 @@ import { defineComponent, ref, watch, computed } from 'vue'
 import { useStore } from '/@/store/index'
 import SelectField from '/@/components/atoms/SelectField.vue'
 import CheckboxInput from '/@/components/atoms/CheckboxInput.vue'
-import SelectButton from '/@/components/atoms/SelectButton.vue'
-import {
-  BoneConstraintName,
-  BoneConstraintOption,
-  getConstraintByName,
-} from '/@/utils/constraints'
-import IKOptionField from '/@/components/molecules/constraints/IKOptionField.vue'
-import LimitRotationOptionField from '/@/components/molecules/constraints/LimitRotationOptionField.vue'
-import CopyLocationOptionField from '/@/components/molecules/constraints/CopyLocationOptionField.vue'
-import CopyRotationOptionField from '/@/components/molecules/constraints/CopyRotationOptionField.vue'
-import UpIcon from '/@/components/atoms/UpIcon.vue'
-import DeleteIcon from '/@/components/atoms/DeleteIcon.vue'
+import { BoneConstraint } from '/@/utils/constraints'
+import ConstraintList from '/@/components/panelContents/ConstraintList.vue'
 import { getBoneIdsWithoutDescendants } from '/@/utils/armatures'
 
 export default defineComponent({
   components: {
     SelectField,
-    SelectButton,
     CheckboxInput,
-    IKOptionField,
-    LimitRotationOptionField,
-    CopyLocationOptionField,
-    CopyRotationOptionField,
-    UpIcon,
-    DeleteIcon,
+    ConstraintList,
   },
   setup() {
     const store = useStore()
@@ -187,65 +119,13 @@ export default defineComponent({
       else draftName.value = ''
     }
 
-    function addConstraint(name: BoneConstraintName) {
+    function updateConstraints(constraints: BoneConstraint[]) {
       if (!lastSelectedBone.value) return
-
-      const constraints = [
-        ...lastSelectedBone.value.constraints,
-        getConstraintByName(name),
-      ]
-      store.updateBone({ constraints })
-    }
-
-    function updateConstraint(index: number, option: BoneConstraintOption) {
-      if (!lastSelectedBone.value) return
-
-      const constraints = lastSelectedBone.value.constraints.concat()
-      constraints.splice(index, 1, { ...constraints[index], option })
-      store.updateBone({ constraints })
-    }
-
-    function deleteConstraint(index: number) {
-      if (!lastSelectedBone.value) return
-
-      const constraints = lastSelectedBone.value.constraints.concat()
-      constraints.splice(index, 1)
-      store.updateBone({ constraints })
-    }
-    function upConstraint(index: number) {
-      if (!lastSelectedBone.value) return
-      if (index === 0) return
-
-      const constraints = lastSelectedBone.value.constraints.concat()
-      const tmp = constraints[index - 1]
-      constraints[index - 1] = constraints[index]
-      constraints[index] = tmp
-      store.updateBone({ constraints })
-    }
-    function downConstraint(index: number) {
-      if (!lastSelectedBone.value) return
-      if (index === lastSelectedBone.value.constraints.length - 1) return
-
-      const constraints = lastSelectedBone.value.constraints.concat()
-      const tmp = constraints[index + 1]
-      constraints[index + 1] = constraints[index]
-      constraints[index] = tmp
       store.updateBone({ constraints })
     }
 
     watch(store.lastSelectedArmature, initDraftName)
     watch(lastSelectedBone, initDraftName)
-
-    const constraintOptions = computed<
-      { value: BoneConstraintName; label: string }[]
-    >(() => {
-      return [
-        { value: 'IK', label: 'IK' },
-        { value: 'LIMIT_ROTATION', label: 'Limit Rotation' },
-        { value: 'COPY_LOCATION', label: 'Copy Location' },
-        { value: 'COPY_ROTATION', label: 'Copy Rotation' },
-      ]
-    })
 
     return {
       draftName,
@@ -285,10 +165,6 @@ export default defineComponent({
           store.updateBone({ parentId: val })
         },
       }),
-      setBoneConstraintName(val: BoneConstraintName) {
-        if (!val) return
-        addConstraint(val)
-      },
       changeArmatureName() {
         if (!draftName.value) return
         if (
@@ -307,11 +183,7 @@ export default defineComponent({
         }
         store.updateBone({ name: draftName.value })
       },
-      updateConstraint,
-      deleteConstraint,
-      upConstraint,
-      downConstraint,
-      constraintOptions,
+      updateConstraints,
     }
   },
 })
@@ -322,6 +194,7 @@ h3 {
   margin-bottom: 10px;
 }
 form {
+  padding-bottom: 10px;
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -342,28 +215,6 @@ form {
       > label + * {
         flex: 1;
         min-width: 50px; // a magic to fix flex width
-      }
-    }
-  }
-  .constraints-item {
-    width: 100%;
-    margin-bottom: 10px;
-    border-top: solid 1px #aaa;
-    .constraint-buttons {
-      display: flex;
-      align-items: center;
-      justify-content: flex-end;
-      > button {
-        margin-left: 8px;
-        border-radius: 8px;
-        width: 18px;
-        height: 18px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        .icon {
-          height: 100%;
-        }
       }
     }
   }
